Use Bun and switch Dev seerve
Bun is better! And the dev server is switched to octane for better peformance in Development. Fixed some migrations cuz they had errors

+++ database/migrations/2020_12_12_102435_support_multiple_docker_images_and_updates.php
@@ -1,5 +1,6 @@
 <?php
 
+
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
@@ -13,8 +14,12 @@ class SupportMultipleDockerImagesAndUpdates extends Migration
   public function up(): void
   {
     Schema::table('eggs', function (Blueprint $table) {
-      $table->json('docker_images')->after('docker_image')->nullable();
-      $table->text('update_url')->after('docker_images')->nullable();
+      if (!Schema::hasColumn('eggs', 'docker_images')) {
+        $table->json('docker_images')->after('docker_image')->nullable();
+      }
+      if (!Schema::hasColumn('eggs', 'update_url')) {
+        $table->text('update_url')->after('docker_images')->nullable();
+      }
     });
 
     switch (DB::getPdo()->getAttribute(PDO::ATTR_DRIVER_NAME)) {
@@ -27,7 +32,9 @@ class SupportMultipleDockerImagesAndUpdates extends Migration
     }
 
     Schema::table('eggs', function (Blueprint $table) {
-      $table->dropColumn('docker_image');
+      if (Schema::hasColumn('eggs', 'docker_image')) {
+        $table->dropColumn('docker_image');
+      }
     });
   }
 
@@ -37,22 +44,28 @@ class SupportMultipleDockerImagesAndUpdates extends Migration
   public function down(): void
   {
     Schema::table('eggs', function (Blueprint $table) {
-      $table->text('docker_image')->after('docker_images');
+      if (!Schema::hasColumn('eggs', 'docker_image')) {
+        $table->text('docker_image')->after('docker_images');
+      }
     });
 
     switch (DB::getPdo()->getAttribute(PDO::ATTR_DRIVER_NAME)) {
       case 'mysql':
-        DB::table('eggs')->update(['docker_images' => DB::raw('JSON_UNQUOTE(JSON_EXTRACT(docker_images, "$[0]")')]);
+        DB::table('eggs')->update(['docker_image' => DB::raw('JSON_UNQUOTE(JSON_EXTRACT(docker_images, "$[0]"))')]);
         break;
       case 'pgsql':
-        DB::table('eggs')->update(['docker_images' => DB::raw('JSON_UNQUOTE(JSON_EXTRACT(docker_images, "$[0]")')]);
-        DB::table('eggs')->update(['docker_images' => DB::raw('docker_images->>0')]);
+        DB::table('eggs')->update(['docker_image' => DB::raw('JSON_UNQUOTE(JSON_EXTRACT(docker_images, "$[0]"))')]);
+        DB::table('eggs')->update(['docker_image' => DB::raw('docker_images->>0')]);
         break;
     }
 
     Schema::table('eggs', function (Blueprint $table) {
-      $table->dropColumn('docker_images');
-      $table->dropColumn('update_url');
+      if (Schema::hasColumn('eggs', 'docker_images')) {
+        $table->dropColumn('docker_images');
+      }
+      if (Schema::hasColumn('eggs', 'update_url')) {
+        $table->dropColumn('update_url');
+      }
     });
   }
 }
